+++ packages/aws-amplify-react/src/Auth/index.jsx
@@ -29,6 +29,7 @@ export { default as ForgotPassword } from './ForgotPassword';
 export { default as Greetings } from './Greetings';
 export { default as FederatedSignIn, FederatedButtons } from './FederatedSignIn';
 export { default as TOTPSetup } from './TOTPSetup';
+export { default as Loading } from './Loading';
 
 export * from './Provider';
 
@@ -62,7 +63,9 @@ export function withAuthenticator(Comp, includeGreetings = false, authenticatorC
                             <Greetings
                                 authState={authState}
                                 authData={authData}
+                                federated={federated || this.props.federated}
                                 onStateChange={this.handleAuthStateChange}
+                                theme={theme}
                             />
                             : null
                         }