+++ packages/aws-amplify/CHANGELOG.md
@@ -3,6 +3,126 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.1.11-unstable.0"></a>
+## [1.1.11-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.10...aws-amplify@1.1.11-unstable.0) (2018-11-13)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.10"></a>
+## [1.1.10](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.10-unstable.6...aws-amplify@1.1.10) (2018-11-12)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.10-unstable.6"></a>
+## [1.1.10-unstable.6](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.10-unstable.5...aws-amplify@1.1.10-unstable.6) (2018-11-12)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.10-unstable.5"></a>
+## [1.1.10-unstable.5](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.10-unstable.4...aws-amplify@1.1.10-unstable.5) (2018-11-10)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.10-unstable.4"></a>
+## [1.1.10-unstable.4](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.10-unstable.3...aws-amplify@1.1.10-unstable.4) (2018-11-09)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.10-unstable.3"></a>
+## [1.1.10-unstable.3](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.10-unstable.2...aws-amplify@1.1.10-unstable.3) (2018-11-09)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.10-unstable.2"></a>
+## [1.1.10-unstable.2](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.10-unstable.1...aws-amplify@1.1.10-unstable.2) (2018-11-09)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.10-unstable.1"></a>
+## [1.1.10-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.10-unstable.0...aws-amplify@1.1.10-unstable.1) (2018-11-06)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.10-unstable.0"></a>
+## [1.1.10-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.9...aws-amplify@1.1.10-unstable.0) (2018-11-06)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.9"></a>
+## [1.1.9](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.9-unstable.1...aws-amplify@1.1.9) (2018-11-01)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.9-unstable.1"></a>
+## [1.1.9-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.9-unstable.0...aws-amplify@1.1.9-unstable.1) (2018-11-01)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.9-unstable.0"></a>
+## [1.1.9-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.8...aws-amplify@1.1.9-unstable.0) (2018-10-30)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.8"></a>
+## [1.1.8](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.8-unstable.3...aws-amplify@1.1.8) (2018-10-29)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.8-unstable.3"></a>
+## [1.1.8-unstable.3](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.8-unstable.2...aws-amplify@1.1.8-unstable.3) (2018-10-29)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
+<a name="1.1.8-unstable.2"></a>
+## [1.1.8-unstable.2](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.8-unstable.1...aws-amplify@1.1.8-unstable.2) (2018-10-29)
+
+
+
+
+**Note:** Version bump only for package aws-amplify
+
 <a name="1.1.8-unstable.1"></a>
 ## [1.1.8-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify@1.1.8-unstable.0...aws-amplify@1.1.8-unstable.1) (2018-10-18)